feat: add private preview badge to browser checks
